requirements.txt
nvgpu==0.10.0
opencv-python==4.8.1.78
pandas==2.1.2
Pillow==10.0.1
pynvml==11.5.0
scipy==1.11.3
seaborn==0.13.0
torch==2.1.0
torchaudio==2.1.0
torchvision==0.16.0
tqdm==4.66.1
triton==2.1.0
tzdata==2023.3
ultralytics==8.0.207
nvidia-pytriton==0.4.0